+++ db/db_impl/db_impl_open.cc
@@ -13,6 +13,7 @@
 #include "db/builder.h"
 #include "db/error_handler.h"
 #include "file/sst_file_manager_impl.h"
+#include "monitoring/persistent_stats_history.h"
 #include "options/options_helper.h"
 #include "rocksdb/wal_filter.h"
 #include "table/block_based/block_based_table_factory.h"
@@ -503,38 +504,101 @@ Status DBImpl::Recover(
   return s;
 }
 
+Status DBImpl::PersistentStatsProcessFormatVersion() {
+  Status s;
+  mutex_.Unlock();
+  if (persistent_stats_cfd_exists_) {
+    // Check persistent stats format version compatibility. Drop and recreate
+    // persistent stats CF if format version is incompatible
+    int sst_format_version = DecodePersistentStatsVersionNumber(
+        this, StatsVersionKeyType::kFormatVersion);
+    int sst_reader_version = DecodePersistentStatsVersionNumber(
+        this, StatsVersionKeyType::kReaderVersion);
+    // abort reading from existing stats CF if any of following is true:
+    // 1. failed to read format version or reader version from disk
+    // 2. sst's format version is below the current minimum supported version
+    // 3. sst's format version is greater than current format version, meaning
+    // this sst is encoded with a newer RocksDB release, and current reader
+    // version is below the sst's reader version
+    if (sst_format_version == -1 || sst_reader_version == -1 ||
+        sst_format_version < kStatsCFMinimumFormatVersion ||
+        (kStatsCFCurrentFormatVersion < sst_format_version &&
+         kPersistentStatsReaderVersion < sst_reader_version)) {
+      ROCKS_LOG_INFO(
+          immutable_db_options_.info_log,
+          "Disable persistent stats due to corrupted or incompatible format "
+          "version\n");
+      DropColumnFamily(persist_stats_cf_handle_);
+      DestroyColumnFamilyHandle(persist_stats_cf_handle_);
+      ColumnFamilyHandle* handle = nullptr;
+      ColumnFamilyOptions co;
+      co.OptimizeForPersistentStats();
+      s = CreateColumnFamily(co, kPersistentStatsColumnFamilyName, &handle);
+      persist_stats_cf_handle_ = static_cast<ColumnFamilyHandleImpl*>(handle);
+    }
+  } else {
+    // Persistent stats CF being created for the first time, need to write
+    // format version key
+    char format_version_key[100];
+    char reader_version_key[100];
+    size_t format_version_key_length = EncodePersistentStatsVersionKey(
+        format_version_key, StatsVersionKeyType::kFormatVersion);
+    size_t reader_version_key_length = EncodePersistentStatsVersionKey(
+        reader_version_key, StatsVersionKeyType::kReaderVersion);
+    if (format_version_key_length == 0 || reader_version_key_length == 0) {
+      ROCKS_LOG_INFO(immutable_db_options_.info_log,
+                     "Failed to encode persistent stats format version\n");
+      s = Status::IOError("Encoding format version failed");
+      return s;
+    }
+    WriteOptions wo;
+    wo.low_pri = true;
+    wo.no_slowdown = true;
+    wo.sync = false;
+    WriteBatch batch;
+    batch.Put(persist_stats_cf_handle_,
+              Slice(format_version_key,
+                    std::min<size_t>(format_version_key_length, 100UL)),
+              ToString(kStatsCFCurrentFormatVersion));
+    batch.Put(persist_stats_cf_handle_,
+              Slice(reader_version_key,
+                    std::min<size_t>(reader_version_key_length, 100UL)),
+              ToString(kPersistentStatsReaderVersion));
+    s = Write(wo, &batch);
+  }
+  mutex_.Lock();
+  return s;
+}
+
 // Initialize the built-in column family for persistent stats. Depending on
 // whether on-disk persistent stats have been enabled before, it may either
 // create a new column family and column family handle or just a column family
 // handle.
 Status DBImpl::InitPersistStatsColumnFamily() {
-  if (persist_stats_cf_handle_) {
-    return Status::OK();
-  }
-  bool persistent_stats_cfd_exists = false;
-  if (versions_->GetColumnFamilySet()->GetColumnFamily(
-          kPersistentStatsColumnFamilyName) != nullptr) {
-    persistent_stats_cfd_exists = true;
-  }
+  assert(!persist_stats_cf_handle_);
+  persistent_stats_cfd_exists_ =
+      versions_->GetColumnFamilySet()->GetColumnFamily(
+          kPersistentStatsColumnFamilyName) != nullptr;
 
-  if (!persistent_stats_cfd_exists) {
-    mutex_.Unlock();
-    ColumnFamilyHandle* handle = nullptr;
-    Status s = CreateColumnFamily(ColumnFamilyOptions(),
-                                  kPersistentStatsColumnFamilyName, &handle);
-    persist_stats_cf_handle_ = static_cast<ColumnFamilyHandleImpl*>(handle);
-    mutex_.Lock();
-    return s;
-  } else {
-    // When recovering from a DB which already contains persistent stats CF,
+  Status s;
+  if (persistent_stats_cfd_exists_) {
+    // We are recovering from a DB which already contains persistent stats CF,
     // the CF is already created in VersionSet::ApplyOneVersionEdit, but
     // column family handle was not. Need to explicitly create handle here.
     persist_stats_cf_handle_ = new ColumnFamilyHandleImpl(
         versions_->GetColumnFamilySet()->GetColumnFamily(
             kPersistentStatsColumnFamilyName),
         this, &mutex_);
-    return Status::OK();
+  } else {
+    mutex_.Unlock();
+    ColumnFamilyHandle* handle = nullptr;
+    ColumnFamilyOptions co;
+    co.OptimizeForPersistentStats();
+    s = CreateColumnFamily(co, kPersistentStatsColumnFamilyName, &handle);
+    persist_stats_cf_handle_ = static_cast<ColumnFamilyHandleImpl*>(handle);
+    mutex_.Lock();
   }
+  return s;
 }
 
 // REQUIRES: log_numbers are sorted in ascending order
@@ -1298,6 +1362,10 @@ Status DBImpl::Open(const DBOptions& db_options, const std::string& dbname,
       impl->DeleteObsoleteFiles();
       s = impl->directories_.GetDbDir()->Fsync();
     }
+  }
+  if (s.ok() && impl->immutable_db_options_.persist_stats_to_disk) {
+    // try to read format version but no need to fail Open() even if it fails
+    s = impl->PersistentStatsProcessFormatVersion();
   }
 
   if (s.ok()) {
