+++ monitoring/in_memory_stats_history.cc
@@ -6,7 +6,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file. See the AUTHORS file for names of contributors.
 
-#include "db/in_memory_stats_history.h"
+#include "monitoring/in_memory_stats_history.h"
 #include "db/db_impl/db_impl.h"
 
 namespace rocksdb {
