+++ monitoring/in_memory_stats_history.h
